feature-policy/header-policy-fullscreen-enabled-for-self-manual.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="f1"></iframe>
<iframe id="f2" allowfullscreen></iframe>
<script>
var srcs = [
  'resources/fullscreen.html',
  '{{location[scheme]}}://{{domains[www]}}:{{location[port]}}/feature-policy/resources/fullscreen.html'
];

function loadFrame(id) {
  var frame = document.getElementById(id);
  for (var src of srcs) {
    var is_fullscreen_enabled = (src === srcs[0] || frame.allowfullscreen);
    promise_test(function(t) {
      frame.src = src;
      return new Promise(function(resolve, reject) {
        window.addEventListener('message', function(e) {
          resolve(e.data);
        }, { once: true });
      }).then(function(data) {
        if (is_fullscreen_enabled) {
          assert_true(data.enabled, 'Document.fullscreenEnabled:');
          assert_equals(data.type, 'change', 'Document.requestFullscreen():');
        } else {
          assert_false(data.enabled, 'Document.fullscreenEnabled:');
          assert_equals(data.type, 'error', 'Document.requestFullscreen():');
        }
      });
  }, (src === srcs[0] ? 'same origin' : 'cross origin') +
    ' iframe with allowfullscreen = ' + frame.allowFullscreen +
    (is_fullscreen_enabled
     ? ' may call document.requestFullscreen() when enabled by feature policy or allowfullscreen attribute'
     : ' may not call document.requestFullscreen() when disabled by feature policy'));
  }
}

loadFrame('f1');
loadFrame('f2');
</script>
